centralize the code to get a `Command::new("cargo")` into a new `user_crate/cargo.rs` file.
