+++ plrust/src/user_crate/build.rs
@@ -6,7 +6,6 @@ All rights reserved.
 Use of this source code is governed by the PostgreSQL license that can be found in the LICENSE.md file.
 */
 
-use std::ffi::CStr;
 use std::{
     path::{Path, PathBuf},
     process::Output,
@@ -14,10 +13,10 @@ use std::{
 
 use color_eyre::{Section, SectionExt};
 use eyre::{eyre, WrapErr};
-use pgx::{pg_sys, PgMemoryContexts};
+use pgx::pg_sys;
 
 use crate::target::{CompilationTarget, CrossCompilationTarget};
-use crate::user_crate::cargo;
+use crate::user_crate::cargo::cargo;
 use crate::{
     gucs,
     user_crate::{CrateState, FnLoad},
@@ -97,56 +96,13 @@ impl FnBuild {
         target_triple: CompilationTarget,
         cross_compilation_target: Option<CrossCompilationTarget>,
     ) -> eyre::Result<(FnLoad, Output)> {
-        let mut command = cargo()?;
+        let mut command = cargo(cargo_target_dir, cross_compilation_target)?;
 
         command.current_dir(&self.crate_dir);
         command.arg("rustc");
         command.arg("--release");
         command.arg("--target");
         command.arg(&target_triple);
-
-        // these are environment variables that could possibly impact pgx compilation that we don't
-        // want to accidentally or purposely be inherited from the running "postgres" process
-
-        command.env_remove("DOCS_RS"); // we'll never be building user function on https://docs.rs
-        command.env_remove("PGX_BUILD_VERBOSE"); // unnecessary to ever build a user function in verbose mode
-        command.env_remove("PGX_PG_SYS_GENERATE_BINDINGS_FOR_RELEASE"); // while an interesting idea, PL/Rust user functions are not used to generate a `pgx` release
-        command.env_remove("CARGO_MANIFEST_DIR"); // we are in the manifest directory b/c of `command.current_dir()` above
-        command.env_remove("OUT_DIR"); // rust's default decision for OUT_DIR is perfectly acceptable to PL/Rust
-        command.env_remove("RUSTC"); // don't let pgx trust the host w.r.t telling it to use a different `rustc`
-
-        command.env("CARGO_TARGET_DIR", &cargo_target_dir);
-        command.env("RUSTFLAGS", "-Clink-args=-Wl,-undefined,dynamic_lookup");
-
-        // There was a time in the past where plrust had a `plrust.pg_config` GUC whose value was
-        // passed down to the "pgx-pg-sys" transient dependency via an environment variable.
-        //
-        // This turned out to be an unwanted bit of user, system, and operational complexity.
-        //
-        // Instead, we tell the environment that "pg_config" is described as environment
-        // variables, and set every property Postgres can tell us (which is essentially how
-        // `pg_config` itself works) as individual environment variables, each prefixed with
-        // "PGX_PG_CONFIG_".
-        //
-        // "pgx-pg-sys"'s build.rs knows how to interpret these environment variables to get what
-        // it needs to properly generate bindings.
-        command.env("PGX_PG_CONFIG_AS_ENV", "true");
-        for (k, v) in pg_config_values() {
-            let k = format!("PGX_PG_CONFIG_{k}");
-            command.env(k, v);
-        }
-
-        // set environment variables we need in order for a cross compile
-        if let Some(target_triple) = cross_compilation_target {
-            // the CARGO_TARGET_xx_LINKER variable
-            let (k, v) = target_triple.linker_envar();
-            command.env(k, v);
-
-            // pgx-specified variable for where the bindings are
-            if let Some((k, v)) = target_triple.bindings_envar() {
-                command.env(k, v);
-            }
-        }
 
         let output = command.output().wrap_err("`cargo` execution failure")?;
 
@@ -224,41 +180,5 @@ impl FnBuild {
     // for #[tracing] purposes
     pub(crate) fn crate_dir(&self) -> &Path {
         &self.crate_dir
-    }
-}
-
-/// Asks Postgres, via FFI, for all of its compile-time configuration data.  This is the full
-/// set of things that Postgres' `pg_config` configuration tool can report.
-///
-/// The returned tuple is a `(key, value)` pair of the configuration name and its value.
-fn pg_config_values() -> impl Iterator<Item = (String, String)> {
-    unsafe {
-        // SAFETY:  we know the memory context we're switching to is valid because we're also making
-        // it right here.  We're also responsible for pfreeing the result of `get_configdata()` and
-        // the easiest way to do that is to simply free an entire memory context at once
-        PgMemoryContexts::new("configdata").switch_to(|_| {
-            let mut nprops = 0;
-            // SAFETY:  `get_configdata` needs to know where the "postmaster" executable is located
-            // and `pg_sys::my_exec_path` is that global, which Postgres assigns once early in its
-            // startup process
-            let configdata = pg_sys::get_configdata(pg_sys::my_exec_path.as_ptr(), &mut nprops);
-
-            // SAFETY:  `get_configdata` will never return the NULL pointer
-            let slice = std::slice::from_raw_parts(configdata, nprops);
-
-            let mut values = Vec::with_capacity(nprops);
-            for e in slice {
-                // SAFETY:  the members (we use) in `ConfigData` are properly allocated char pointers,
-                // done so by the `get_configdata()` call above
-                let name = CStr::from_ptr(e.name);
-                let setting = CStr::from_ptr(e.setting);
-                values.push((
-                    name.to_string_lossy().to_string(),
-                    setting.to_string_lossy().to_string(),
-                ))
-            }
-
-            values.into_iter()
-        })
     }
 }
